+++ webpack.config.js
@@ -35,8 +35,8 @@ let webpackConfig = {
             filename: "[name].min.css"
         }),
         new HtmlWebpackPlugin({
-            template: 'pages/event-list.html',
-            filename: 'event-list.html',
+            template: 'pages/schedule.html',
+            filename: 'schedule.html',
             inject: 'head',
             config: config.options
         }),